Move Sha256Hasher to storage module
